splitting db tests
