+++ package.json
@@ -11,7 +11,7 @@
     "dev:migrate": "NODE_ENV=development NODE_OPTIONS='--import tsx' npx knex migrate:latest",
     "dev:start": "NODE_ENV=development NODE_OPTIONS='--import tsx' nodemon -q ./src/index.ts",
     "test:unit": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/unit/**/*.test.ts",
-    "test:integration": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/integration/*.test.ts",
+    "test:integration": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/integration/db/*.test.ts",
     "test": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/**/*.test.ts",
     "test:migrate": "NODE_ENV=test NODE_OPTIONS='--import tsx' npx knex migrate:latest",
     "test:seed": "NODE_ENV=test NODE_OPTIONS='--import tsx' npx knex seed:run",
